skins: Fix Skin::buildSidebar to not share cache between skins
The output of the various overridable methods called here, as well as
the onSkinBuildSidebar() hook, may vary by skin. As such, the cache
must take this into account.

The reason this has likely not been an issue with the majority of third
party wikis that vary behaviour between skins, is that by configuration
$wgEnableSidebarCache is disabled by default.

As such, in addition to the documented responsiblity of site operators
to know whether they have extensions that vary sidebar by title
or user (which is what $wgEnableSidebarCache requires), they also had
to know that it doesn't vary by skin, which seems unreasonable and
simply isn't needed given the low cardinality of skins.

Bug: T303007
Change-Id: I0ba41eff550a5ed46e433d068a103f6ea24e5b15